refactor: Create common tests directory and migrate method_registry tests
Create src/common/tests/ directory structure and extract inline test
module from method_registry.rs into dedicated test file. Maintains all
22 tests with identical behavior.
