+++ src/common/mod.rs
@@ -12,6 +12,9 @@ pub mod error_renderer;
 pub mod string_similarity;
 pub mod method_registry;
 
+#[cfg(test)]
+mod tests;
+
 // Forward declare VirtualMachine for NativeFn signature
 // We can't import VirtualMachine directly as it would create a circular dependency
 // The actual implementation will be in vm/mod.rs
