+++ src/common/tests/mod.rs
@@ -1,0 +1,1 @@
+mod method_registry;
